move files to the modal they are used. Only load them locally
